requirements.txt
anyio==3.6.2
async-timeout==4.0.2
boto3==1.26.84
botocore==1.29.84
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
fastapi==0.92.0
faunadb==4.5.0
future==0.18.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
iso8601==1.1.0
jmespath==1.0.1
pydantic==1.10.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
redis==4.5.1
requests==2.28.2
rfc3986==1.5.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
